Move Scala 2 only code to specific directory
